requirements.txt
beautifulsoup4>=4.12.0
markdownify>=0.12.0
lxml>=4.9.0
playwright>=1.45.0
pdfminer.six>=20231228
PyPDF2>=3.0.0
